portal/server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pymysql>=1.1.0
anthropic>=0.18.0
claude-agent-sdk>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-pptx>=0.6.21
python-docx>=1.1.0
Pillow>=10.0.0
PyYAML>=6.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0
minio>=7.2.0
aiofiles>=23.0.0
